+++ review_service/src/main/java/org/personal/review_service/repository/ReviewRepositoryCustom.java
@@ -1,0 +1,12 @@
+package org.personal.review_service.repository;
+
+import org.personal.review_service.domain.Review;
+
+import java.util.List;
+
+public interface ReviewRepositoryCustom {
+
+    List<Review> findReviewsByUserId(Long userId);
+
+    List<Review> findReviewsByLocationId(Long locationId);
+}
