+++ review_service/src/main/java/org/personal/review_service/repository/ReviewRepositoryImpl.java
@@ -1,0 +1,40 @@
+package org.personal.review_service.repository;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import org.personal.review_service.domain.QReview;
+import org.personal.review_service.domain.Review;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public class ReviewRepositoryImpl implements ReviewRepositoryCustom{
+
+    private JPAQueryFactory jpaQueryFactory;
+
+    @Autowired
+    public ReviewRepositoryImpl(JPAQueryFactory jpaQueryFactory) {
+        this.jpaQueryFactory = jpaQueryFactory;
+    }
+
+    @Override
+    public List<Review> findReviewsByUserId(Long userId) {
+        QReview review = QReview.review;
+
+        return jpaQueryFactory
+                .selectFrom(review)
+                .where(review.userId.eq(userId))
+                .fetch();
+    }
+
+    @Override
+    public List<Review> findReviewsByLocationId(Long locationId) {
+        QReview review = QReview.review;
+
+        return jpaQueryFactory
+                .selectFrom(review)
+                .where(review.locationId.eq(locationId))
+                .fetch();
+    }
+}
